+++ src/main/java/com/chzzkGamble/gamble/DonationHandler.java
@@ -1,0 +1,38 @@
+package com.chzzkGamble.gamble;
+
+import com.chzzkGamble.chzzk.dto.DonationMessage;
+import com.chzzkGamble.event.DonationEvent;
+import com.chzzkGamble.utils.StringParser;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.event.EventListener;
+import org.springframework.scheduling.annotation.Async;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class DonationHandler {
+
+    private static final char LEFT_DELIMITER = '<';
+    private static final char RIGHT_DELIMITER = '>';
+
+    private final GambleService gambleService;
+
+    @Async
+    @EventListener(DonationEvent.class)
+    public void voteGamble(DonationEvent donationEvent) {
+        DonationMessage donationMessage = (DonationMessage) donationEvent.getSource();
+        String elementName = getElementName(donationMessage.getMsg());
+        if (elementName == null) {
+            return; //룰렛용 도네가 아님
+        }
+        gambleService.vote(donationMessage.getChannelName(), elementName, donationMessage.getCheese());
+    }
+
+    private String getElementName(String msg) {
+        try {
+            return StringParser.parseFromTo(msg, LEFT_DELIMITER, RIGHT_DELIMITER);
+        } catch (IllegalArgumentException e) {
+            return null;
+        }
+    }
+}
