+++ src/main/java/com/chzzkGamble/gamble/GambleService.java
@@ -1,0 +1,6 @@
+package com.chzzkGamble.gamble;
+
+public interface GambleService {
+
+    void vote(String channelName, String elementName, int cheese);
+}
